Adapt to unicode refactoring
