+++ users/spidey3/rules.mk
@@ -11,5 +11,5 @@ ifeq ($(strip $(RGBLIGHT_ENABLE)), yes)
   SRC += layer_rgb.c
 endif
 ifeq ($(strip $(UNICODEMAP_ENABLE)), yes)
-  SRC += unicode.c
+  SRC += spidey3_unicode.c
 endif
